requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
alembic
psycopg2-binary
httpx
python-multipart
opencv-python
faiss-cpu
pytesseract
apscheduler
PyJWT
pydantic[email]
python-dotenv
